allow overlay to take precedence over main tree
Since we provide specifically patched versions of
some stuff to better support Apple Silicon, it's
a good idea for our ebuilds to be the default ones
installed when the user does not specify a repo.
Advanced users are of course able to override this
behaviour by appending "::gentoo" to the package
atom.

Signed-off-by: James Calligeros <jcalligeros99@gmail.com>