asahi-sources: safely install kernel and display messages
Previously, the install phase would result in a Portage
Sandbox violation. We fix this by calling dodir and using
the directory imaging feature provided by Portage rather
than trying to access the live filesystem ourselves.

For messages to be displayed at the end of emerging
when --jobs or --quiet are set, we use elog instead
of einfo in the post{inst,rm} functions. This is in
line with the kernel-2 eclass used by the main Gentoo
tree.

Signed-off-by: James Calligeros <jcalligeros99@gmail.com>