sys-apps/asahi-scripts: adjust dracut config to match kernel config
The new kernel config we're proposing has
implications for how dracut needs to be configured,
as we build more of the Asahi drivers directly into
the kernel than the previous configuration did.

We also explicitly tell grub-mkimage to use
xz compression, as I see that chadmed has been
concerned about the size of the LiveISO in the past.

Signed-off-by: zzywysm <zzy@zzywysm.com>