sys-apps/asahi-scripts: only install dracut modules for livecd
We need to coexist peacefully on the arm64 livecd. Other platforms
do not want our hostonly assumptions which strip them of their drivers
and such.

Introduce the livecd USE flag, which will only install the raw dracut
modules and nothing else. This lets us specify them manually in the
Catalyst build process for the asahi-kernel build.

Signed-off-by: James Calligeros <jcalligeros99@gmail.com>