sys-kernel/asahi-kernel: add 6.16.4_p2
Base asahi-kernel on gentoo-kernel-6.16.4.ebuild. This replaces
genpatches with Gentoo's distro kernel patchset. This has the advantage
that it does not need to be updated for each stable release. Only
downside is that it is no longer possible to have non-stable base
kernels. This should be fine for dist-kernel at this stage.

The ebuild copies support for gentoo-kernel's architectures and features
to minimise the diff to gentoo-kernel-6.16.4.ebuild. This should be dead
code and not cause any trouble.

linux-6.16.4_apple_silicon_base.config is the lightly edited result of
```
grep \
  -e _APPLE \
  -e K_PAGES \
  -e CONFIG_RUST= \
  -e _ASAHI \
  -e _MACSMC \
  -e _BRCMFMAC= \
  -e _BRCMFMAC_PCIE \
  -e _BCM4377 \
  -e _MAGICMOUSE \
  -e _RTKIT \
  -e _ADP= \
  -e _AOP \
  -e _CS42L8 \
  -e '_TAS27[67]' \
  -e _XHCI_PCI_ASMEDIA \
  -e _ARM64_MEMORY_MODEL \
  -e _ARM64_ACTLR \
  config-6.16.4-401.asahi.fc42.aarch64+16k
```

Signed-off-by: Janne Grunau <j@jannau.net>